refactor(mailbridge): split runtime classes and add orchestration-backed unit tests
