+++ docs/features/active/2026-04-05-refactor-and-test-9/code-review.2026-04-06T14-58.md
@@ -1,0 +1,9 @@
+# Code Review — 2026-04-06T14:58Z
+
+## Findings
+- Runtime classes were split from monolithic `Program.cs` into dedicated production files.
+- Internal visibility and test project references were added to enable runtime unit tests.
+- New runtime-focused test suite validates startup parsing/loading, state store transitions, scanner waiting behavior, cache repository persistence, RPC handlers, oversized response downgrade, and COM lookup fallback.
+
+## Risks
+- Several Windows/COM-sensitive runtime files remain under-covered on Linux execution.
