+++ docs/features/active/2026-04-05-refactor-and-test-9/evidence/baseline/policy-read.2026-04-06T14-55.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-06T14:55:00Z
+Command: find .github -maxdepth 3 -type f | sort
+EXIT_CODE: 0
+Output Summary: `.github` directory is not present in this checkout, so policy-compliance-order documents were not available to read; proceeded with best-effort and documented the gap.
