+++ docs/features/active/2026-04-05-refactor-and-test-9/evidence/qa-gates/coverage-summary.2026-04-06T14-56.md
@@ -1,0 +1,12 @@
+Timestamp: 2026-04-06T14:56:30Z
+Command: python parse of cobertura report
+EXIT_CODE: 0
+Output Summary:
+- OpenClaw.MailBridge/BridgeStateStore.cs: 100.0%
+- OpenClaw.MailBridge/CacheRepository.cs: 100.0%
+- OpenClaw.MailBridge/ComActiveObject.cs: 77.8%
+- OpenClaw.MailBridge/OutlookScanner.cs: 35.0%
+- OpenClaw.MailBridge/OutlookStaExecutor.cs: 25.8%
+- OpenClaw.MailBridge/PipeRpcWorker.cs: 26.2%
+- OpenClaw.MailBridge/Program.cs: 37.7%
+- OpenClaw.MailBridge/ScanWorker.cs: 0.0%
