+++ docs/features/active/2026-04-05-refactor-and-test-9/evidence/qa-gates/coverage.2026-04-06T14-55.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-06T14:56:00Z
+Command: /root/.dotnet/dotnet test tests/OpenClaw.MailBridge.Tests/OpenClaw.MailBridge.Tests.csproj --collect:"XPlat Code Coverage" -v minimal
+EXIT_CODE: 0
+Output Summary: Coverage collector produced `coverage.cobertura.xml`; per-file parsing shows some refactored runtime files below 80% in this Linux execution environment (especially Windows/COM-sensitive classes).
