+++ docs/features/active/2026-04-05-refactor-and-test-9/evidence/qa-gates/dotnet-test.2026-04-06T14-55.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-06T14:55:30Z
+Command: /root/.dotnet/dotnet test OpenClaw.MailBridge.sln -v minimal
+EXIT_CODE: 0
+Output Summary: All solution tests passed (18/18).
