+++ docs/features/active/2026-04-05-refactor-and-test-9/feature-audit.2026-04-06T14-58.md
@@ -1,0 +1,12 @@
+# Feature Audit — 2026-04-06T14:58Z
+
+## Acceptance Criteria Status
+- Source: `docs/features/active/2026-04-05-refactor-and-test-9/issue.md`
+- Total AC items: 5
+- Checked off (delivered): 4
+- Remaining (unchecked): 1
+- Items remaining:
+  - Coverage evidence is produced showing at least 80% line coverage for each newly tested runtime file.
+
+## Gate
+- Overall: **not fully ready** (coverage-per-file target unmet in current environment).
