+++ docs/features/active/2026-04-05-refactor-and-test-9/issue.md
@@ -1,0 +1,21 @@
+# refactor-and-test (Issue #9)
+
+- Date captured: 2026-04-05
+- Author: codex
+- Status: Active
+
+- Issue: #9
+- Last Updated: 2026-04-06
+- Work Mode: full-feature
+
+## Summary
+
+Refactor `OpenClaw.MailBridge` so each runtime class has a dedicated production file and add policy-compliant unit tests to improve maintainability and confidence.
+
+## Acceptance Criteria
+
+- [x] Runtime classes currently co-located in `src/OpenClaw.MailBridge/Program.cs` are split so each class has its own dedicated production file.
+- [x] Existing MailBridge runtime behavior remains intact after the refactor (no API contract regressions for bridge startup and internal runtime components).
+- [x] Unit tests are added for the refactored runtime files in `src/OpenClaw.MailBridge`.
+- [ ] Coverage evidence is produced showing at least 80% line coverage for each newly tested runtime file.
+- [x] Solution tests pass after refactor and test additions.
