+++ docs/features/active/2026-04-05-refactor-and-test-9/plan.2026-04-06T14-25.md
@@ -1,0 +1,17 @@
+### Phase 0 — Intake, Policy, and Baseline
+- [x] [P0-T1] Read applicable repository policy instructions and collect baseline repo status evidence.
+- [x] [P0-T2] Confirm request classification and change-budget route selection.
+- [x] [P0-T3] Initialize orchestration checkpoint and active feature artifacts.
+
+### Phase 1 — Refactor Runtime Classes
+- [x] [P1-T1] Split runtime classes from `src/OpenClaw.MailBridge/Program.cs` into dedicated production files.
+- [x] [P1-T2] Keep runtime wiring behavior and contracts unchanged after refactor.
+
+### Phase 2 — Unit Test Expansion
+- [x] [P2-T1] Add project references/internals access needed for runtime class tests.
+- [x] [P2-T2] Add unit tests that exercise each refactored runtime file/class.
+
+### Phase 3 — QA and Review
+- [x] [P3-T1] Run formatting/build/tests and record results.
+- [x] [P3-T2] Generate coverage report and verify per-file coverage target.
+- [x] [P3-T3] Update acceptance criteria checklist and orchestration checkpoint with completion summary.
