+++ docs/features/active/2026-04-05-refactor-and-test-9/policy-audit.2026-04-06T14-58.md
@@ -1,0 +1,5 @@
+# Policy Audit — 2026-04-06T14:58Z
+
+- `.github` policy instruction files required by `policy-compliance-order` were not present in this checkout.
+- Proceeded with best-effort policy adherence and documented the missing-policy dependency gap.
+- No secrets or `.env` files were introduced.
