+++ docs/features/active/2026-04-05-refactor-and-test-9/user-story.md
@@ -1,0 +1,10 @@
+# User Story: refactor-and-test
+
+As a maintainer of OpenClaw MailBridge,
+I want runtime classes separated into dedicated files with high unit-test coverage,
+so that code ownership, reviewability, and regression detection are improved.
+
+## Acceptance
+- Refactor complete with one runtime class per production file.
+- Unit coverage demonstrates 80%+ per targeted file.
+- End-to-end tests remain green.
