+++ src/OpenClaw.MailBridge/OpenClaw.MailBridge.csproj
@@ -10,6 +10,11 @@
     <ProjectReference Include="..\OpenClaw.MailBridge.Contracts\OpenClaw.MailBridge.Contracts.csproj" />
   </ItemGroup>
   <ItemGroup>
+    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+      <_Parameter1>OpenClaw.MailBridge.Tests</_Parameter1>
+    </AssemblyAttribute>
+  </ItemGroup>
+  <ItemGroup>
     <PackageReference Include="Microsoft.Data.Sqlite" Version="8.0.11" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="8.0.1" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.1" />
