+++ tests/OpenClaw.MailBridge.Tests/OpenClaw.MailBridge.Tests.csproj
@@ -11,8 +11,10 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
     <PackageReference Include="MSTest.TestAdapter" Version="3.6.4" />
     <PackageReference Include="MSTest.TestFramework" Version="3.6.4" />
+    <PackageReference Include="coverlet.collector" Version="6.0.2" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\OpenClaw.MailBridge.Contracts\OpenClaw.MailBridge.Contracts.csproj" />
+    <ProjectReference Include="..\..\src\OpenClaw.MailBridge\OpenClaw.MailBridge.csproj" />
   </ItemGroup>
 </Project>
